Implement polling manager for periodic api calls
- issue periodic api call to get appliances in main vm
- handle session lost api response:
* add requires session marker interface for api calls that need a valid session
* add session checker ktor custom plugin
* show alert dialog + navigate to welcome
- refactor restore session use case -> implement silent login api call
- add infra package and koin module for application/util type classes
- move UnauthenticatedException to data module
- change type of userId login parameter to Long
- refactor code and package structure

+++ composeApp/src/commonMain/kotlin/com/teoryul/mytesy/data/api/model/login/LoginResponse.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class LoginResponse(
-    val userID: Int? = null,
+    val userID: Long? = null,
     val password: String? = null,
     val email: String? = null,
     val firstName: String? = null,
